requirements.txt
firebase-admin==6.7.0
streamlit==1.44.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.166.0
imapclient==3.0.1
dotenv-python==0.0.1

google-generativeai==0.8.5
prometheus_client==0.21.1
html2text==2025.4.15


